some improovements

+++ PostService.js
@@ -1,36 +1,46 @@
 import Post from "./Post.js";
 
 class PostService {
-  async createPost(post) {
+  async createPost(post, picture) {
     const createdPost = await Post.create(post);
     return createdPost;
   }
 
   async getAll() {
-      const posts = await Post.find();
-      return posts
+    const posts = await Post.find();
+    return posts;
+  }
+
+  async getOnePost(id) {
+    if (!id) {
+      throw new Error("NO ID");
     }
 
-    async getOnePost(id) {
-      
-        const post = await Post.findById(id);
-      
-        return post
-    }
-    async updatePost(post) {
-      const updatedPost = await Post.findByIdAndUpdate(post._id, post, {
-        new: true,
-      });
-        return updatedPost;
-    }
-     
-    async deletePost(id) {
+    const post = await Post.findById(id);
 
-        const deletedPost = await Post.findByIdAndDelete(id);
-
-        return deletedPost
-        
+    return post;
+  }
+  async updatePost(post) {
+    if (!post._id) {
+      throw new Error("NO ID");
     }
+
+    const updatedPost = await Post.findByIdAndUpdate(post._id, post, {
+      new: true,
+    });
+
+    return updatedPost;
+  }
+
+  async deletePost(id) {
+    if (!id) {
+      throw new Error("WRONG ID");
+    }
+
+    const deletedPost = await Post.findByIdAndDelete(id);
+
+    return deletedPost;
+  }
 }
 
-export default new PostService()
+export default new PostService();
